Gas Optimised contracts

+++ src/4_Lock.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.13;
 
 contract Lock {
-    uint8 locked;
+    uint locked;
 
     function isLocked() public view returns (bool) {
         return locked == 0;
